+++ assets/jsx/workflow-editor/components/data-fields/conditional/components/field-expression-builder.jsx
@@ -1,8 +1,7 @@
 import { ConditionalExpressionBuilder } from './conditional-expression-builder';
 
-export const FieldExpressionBuilder = ({ options, value, handleOnChange, context }) => {
+export const FieldExpressionBuilder = ({ value, handleOnChange, context }) => {
     return <ConditionalExpressionBuilder
-        options={options}
         value={value}
         handleOnChange={handleOnChange}
         context={context}