+++ assets/jsx/workflow-editor/components/data-fields/conditional/components/value-expression-builder.jsx
@@ -2,16 +2,14 @@ import { useMemo } from '@wordpress/element';
 import { getVariableDataTypeByVariableName } from '../../../../utils';
 import { ConditionalExpressionBuilder } from './conditional-expression-builder';
 
-export const ValueExpressionBuilder = ({ options, value, handleOnChange, context, rule }) => {
+export const ValueExpressionBuilder = ({ value, handleOnChange, context, rule }) => {
+    // Get the variable data type
     const variableDataType = useMemo(
         () => getVariableDataTypeByVariableName(rule.field, context.variables),
         [rule.field, context.variables]
     );
 
-    console.log(variableDataType);
-
     return <ConditionalExpressionBuilder
-        options={options}
         value={value}
         handleOnChange={handleOnChange}
         context={context}