+++ assets/jsx/workflow-editor/components/data-fields/post-query-conditional/index.jsx
@@ -1,0 +1,3 @@
+import PostQueryConditional from './components/post-query-conditional';
+
+export default PostQueryConditional;