+++ src/Modules/Workflows/Domain/Steps/Actions/Definitions/Conditional.php
@@ -104,6 +104,11 @@ class Conditional implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [];
+    }
+
     public function getOutputSchema(): array
     {
         return [