+++ src/Modules/Workflows/Domain/Steps/Actions/Definitions/QueryPosts.php
@@ -128,6 +128,11 @@ class QueryPosts implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [];
+    }
+
     public function getOutputSchema(): array
     {
         return [