+++ src/Modules/Workflows/Domain/Steps/Actions/Definitions/RemovePostTerm.php
@@ -132,6 +132,11 @@ class RemovePostTerm implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [];
+    }
+
     public function getOutputSchema(): array
     {
         return [