+++ src/Modules/Workflows/Domain/Steps/Actions/Definitions/SendRay.php
@@ -113,6 +113,11 @@ class SendRay implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [];
+    }
+
     public function getOutputSchema(): array
     {
         return [