+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnAdminInit.php
@@ -75,6 +75,11 @@ class OnAdminInit implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [];
+    }
+
     public function getOutputSchema(): array
     {
         return [];