+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostMetaChange.php
@@ -135,6 +135,13 @@ class OnPostMetaChange implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+
+        ];
+    }
+
     public function getOutputSchema(): array
     {
         return [