+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostRowAction.php
@@ -159,6 +159,13 @@ class OnPostRowAction implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+
+        ];
+    }
+
     public function getOutputSchema(): array
     {
         return [