+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostSchedule.php
@@ -122,6 +122,13 @@ class OnPostSchedule implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+
+        ];
+    }
+
     public function getOutputSchema(): array
     {
         return [