+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostStatusChange.php
@@ -123,6 +123,13 @@ class OnPostStatusChange implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+
+        ];
+    }
+
     public function getOutputSchema(): array
     {
         return [