+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostUpdate.php
@@ -61,7 +61,7 @@ class OnPostUpdate implements StepTypeInterface
     {
         return [
             [
-                "label" => __("Post Query", "post-expirator"),
+                "label" => __("Conditions", "post-expirator"),
                 "description" => __(
                     "Specify the criteria for posts that will trigger this action.",
                     "post-expirator"
@@ -69,19 +69,12 @@ class OnPostUpdate implements StepTypeInterface
                 "fields" => [
                     [
                         "name" => "postQuery",
-                        "type" => "postQuery",
+                        "type" => "postQueryConditional",
                         "label" => __("Post query", "post-expirator"),
                         "description" => __(
                             "The query defines the posts that will trigger this action.",
                             "post-expirator"
                         ),
-                        "settings" => [
-                            "acceptsInput" => false,
-                            "isPostTypeRequired" => true,
-                            "postTypeDescription" => __("Select the post types that will trigger this action.", "post-expirator"),
-                            "postIdDescription" => __("Enter one or more post IDs. Leave empty to include all posts.", "post-expirator"),
-                            "postStatusDescription" => __("If selected, only posts with these statuses will trigger this action.", "post-expirator"),
-                        ],
                         "default" => [
                             "postSource" => "custom",
                             "postType" => ["post"],
@@ -118,6 +111,24 @@ class OnPostUpdate implements StepTypeInterface
                         "rule" => "hasOutgoingConnection",
                     ],
                 ]
+            ]
+        ];
+    }
+
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+            [
+                "name" => "postBefore",
+                "type" => "post",
+                "label" => __("Post Before Update", "post-expirator"),
+                "description" => __("The post that was saved, with the old properties.", "post-expirator"),
+            ],
+            [
+                "name" => "postAfter",
+                "type" => "post",
+                "label" => __("Post After Update", "post-expirator"),
+                "description" => __("The post that was saved, with the new properties.", "post-expirator"),
             ]
         ];
     }