+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnSchedule.php
@@ -95,6 +95,13 @@ class OnSchedule implements StepTypeInterface
         ];
     }
 
+    public function getStepScopedVariablesSchema(): array
+    {
+        return [
+
+        ];
+    }
+
     public function getOutputSchema(): array
     {
         return [];